JWT support - WIP - test fixes - refactored to services for mock testing
